makefile for project
Makefile allows for anyone downloading the code to use terminal and type the command "make" and then the entirety of the compilation will be done based on the makefile.  Then to run the program the user types "./carTrip"  (or however you wish to rename the variable PROG). Finally, to clean up all the new files created before say, a push, you can use the command "make clean" to delete all object files and the .carTrip file.  As a dev you probably won't notice the need for it as the code is set up in your CodeBlocks (or whichever program you're using).  However, as a user downloading the code for a few runs it doesn't make sense to set up the project in code blocks so all I have to do is use this makefile.